Project Restructure
Moved VKR library files into VKR directory, and added support for RelWithDbgInfo cmake configuration.
