+++ VKR/include/VKR/VKR.h
@@ -13,6 +13,24 @@
 #define __PRETTY_FUNCTION__ __FUNCSIG__ 
 #endif
 
+#ifndef VKR_DEBUG
+#define EASY_FUNCTION(...)
+#define EASY_BLOCK(...)
+#define EASY_END_BLOCK
+#define EASY_THREAD_SCOPE(...)
+#define EASY_PROFILER_ENABLE
+#define EASY_MAIN_THREAD
+#define PROFILER_START_LISTENING
+#define PROFILER_STOP_LISTENING
+#define PROFILER_DUMP(filename)
+#else
+#define PROFILER_START_LISTENING profiler::startListen()
+#define PROFILER_STOP_LISTENING profiler::stopListen()
+#if VKR_PROFILER_RECORD
+#define PROFILER_DUMP(filename) profiler::dumpBlocksToFile(filename)
+#endif
+#endif
+
 namespace VKR {
     /**
      * @brief Initializes VKR.
