+++ VKR/src/File.cpp
@@ -1,18 +1,7 @@
-#include "../include/File.h"
-#include "../include/Logger.h"
+#include "../include/VKR/File.h"
+#include "../include/VKR/Logger.h"
 #include <fstream>
 #include <easy/profiler.h>
-
-/*
-#if defined(_WIN32)
-#include <windows.h>
-#else
-#define MAX_PATH 65536
-#include <stdlib.h>
-#include <sys/stat.h>
-#include <unistd.h>
-#endif
-*/
 
 const bool VKR::IO::FileExists(const char* filePath)
 {
